+++ src/main/java/com/udea/proyectofinal/service/ProductoService.java
@@ -4,6 +4,7 @@ import com.udea.proyectofinal.model.Cliente;
 import com.udea.proyectofinal.model.Producto;
 import com.udea.proyectofinal.repository.ClienteRepository;
 import com.udea.proyectofinal.repository.ProductoRepository;
+import com.udea.proyectofinal.repository.VentaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +14,9 @@ import java.util.List;
 public class ProductoService {
     @Autowired
     private ProductoRepository productoRepository;
+
+    @Autowired
+    private VentaRepository ventaRepository;
 
     //Todos los productos
     public List<Producto> getAllProducto() {
@@ -31,6 +35,11 @@ public class ProductoService {
 
     // borrar producto
     public void deleteProducto(Long id) {
+        // Verificar si hay ventas asociadas al producto
+        long ventasCount = ventaRepository.countByProductoId(id);
+        if (ventasCount > 0) {
+            throw new RuntimeException("No se puede eliminar el producto porque tiene " + ventasCount + " venta(s) asociada(s)");
+        }
         productoRepository.deleteById(id);
     }
